move CdsRetryPolicy to inner class of CdsProfileFetcher. removed log message.
